// category/src/modules/grpc/proto/v2.pb.ts
export interface GetVariantsRequest {
  categoryId: string;
}