// category/src/modules/grpc/proto/category.pb.ts
export interface GetCatConPriceRangeRequest {
  variantId: string;
  conditionId: string;
  catConditionQuality: string;
}

export interface GetCatConPriceRangeResponse {
  priceRange: PriceNudge | undefined;
}

export interface PriceNudge {
  min: number;
  max: number;
}

export interface GetProductCatConRequest {
  id: string;
  variantId: string;
  sellPrice: number;
}